requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.7.1

pymongo>=4.5.0
mongoengine>=0.27.0
motor>=3.3.0               

langchain>=0.3.27
langchain-core>=0.3.75
langchain-groq>=0.3.7
llama-index>=0.8.0
PyMuPDF>=1.24.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
pdf2image                 
Pillow>=10.0.0            
pytesseract               

fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

pydantic>=2.0.0               
python-dotenv>=1.0.0
python-dateutil>=2.8.2        
requests>=2.31.0
tqdm>=4.65.0                  
python-multipart              
aiofiles                      
asyncio-throttle              
tenacity>=8.2.0               

cryptography>=41.0.0          
python-magic                  
bleach                        
validators                    

aiosmtplib                    
aioimaplib                    

plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

black>=23.0.0           
flake8>=6.0.0           